add entity class, shield system
