+++ src/character.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "common.hpp"
+#include "entity.hpp"
 #include "potion.hpp"
 #include <entt/entity/registry.hpp>
 #include <vector>
