+++ src/components/physicsScaleComponent.hpp
@@ -1,0 +1,7 @@
+#pragma once
+#include <common.hpp>
+
+struct physicsScaleComponent
+{
+    vec2 scale;
+};
