+++ src/factory.cpp
@@ -18,12 +18,14 @@ entt::entity makeCharacter(entt::registry& registry)
 entt::entity makeShield(entt::registry& registry)
 {
     const entt::entity entity = registry.create();
-    registry.assign<shieldComponent>(entity);
+    registry.assign<shieldComponent>(entity, false);
     auto& mo = registry.assign<motionComponent>(entity);
-    // Setting initial values
+    // Setting initial values3
     mo.position = { 50.f, 300.f };
     mo.radians = 20.f;
     mo.speed = 200.f;
     registry.assign<inputMouse>(entity, 0.0, 0.0);
+    auto &ps = registry.assign<physicsScaleComponent>(entity);
+    ps.scale = { 0.5f, 0.05f };
     return entity;
 }
