+++ src/pebbles.hpp
@@ -3,6 +3,7 @@
 #include <vector>
 
 #include "common.hpp"
+#include "entity.hpp"
 
 // !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 // DON'T WORRY ABOUT THIS CLASS UNTIL ASSIGNMENT 3
