+++ src/projectile.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "common.hpp"
+#include "entity.hpp"
 
 // Salmon enemy
 class Projectile : public Entity {
