+++ src/water.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "common.hpp"
+#include "entity.hpp"
 
 class Water : public Entity {
 public:
